+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Change log
 
-## Version 0.1.0
+## Version 0.1.0 
 
 - Fix documentation
 - Remove ConstantRef
 - Add better method and field descriptions 
-- Add PoolAccess Monad
+- Introducing a stageing system
+- Added Code writing
+- Changed Megaparsec to Attoparsec of improved performance.
+
+- Add Attributes 
+  - LineNumberTable
 
 ## Version 0.0.2